Created separate signups for different user tpyes
- Sign-up page leads to a page inquiring the user what type of account they want to make
- Buttons lead to specific sign-up page
