+++ src/app/auth/signup/page.tsx
@@ -1,127 +1,39 @@
+/* eslint-disable max-len */
+
 'use client';
 
-import { signIn } from 'next-auth/react';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import * as Yup from 'yup';
-import { Card, Col, Container, Button, Form, Row } from 'react-bootstrap';
-import { createUser } from '@/lib/dbActions';
-
-type SignUpForm = {
-  email: string;
-  password: string;
-  confirmPassword: string;
-  // role: Role;
-  // acceptTerms: boolean;
-};
+import { Container, Col, Button, Row, Card } from 'react-bootstrap';
 
 /** The sign up page. */
-const SignUp = () => {
-  const validationSchema = Yup.object().shape({
-    email: Yup.string().required('Email is required').email('Email is invalid'),
-    password: Yup.string()
-      .required('Password is required')
-      .min(6, 'Password must be at least 6 characters')
-      .max(40, 'Password must not exceed 40 characters'),
-    confirmPassword: Yup.string()
-      .required('Confirm Password is required')
-      .oneOf([Yup.ref('password'), ''], 'Confirm Password does not match'),
-  });
-
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors },
-  } = useForm<SignUpForm>({
-    resolver: yupResolver(validationSchema),
-  });
-
-  const onSubmit = async (data: SignUpForm) => {
-    // console.log(JSON.stringify(data, null, 2));
-    await createUser(data);
-    // After creating, signIn with redirect to the add page
-    await signIn('credentials', { callbackUrl: '/add', ...data });
-  };
-
-  return (
-    <main>
-      <Container className="text-font">
-        <Row className="justify-content-center">
-          <Col xs={5}>
-            <h1 className="text-center">Sign Up</h1>
-            <Card>
-              <Card.Body>
-                <Form onSubmit={handleSubmit(onSubmit)}>
-                  <Form.Group className="form-group">
-                    <Form.Label>Email</Form.Label>
-                    <input
-                      type="text"
-                      {...register('email')}
-                      className={`form-control ${errors.email ? 'is-invalid' : ''}`}
-                    />
-                    <div className="invalid-feedback">{errors.email?.message}</div>
-                  </Form.Group>
-
-                  {/* For users to create an account type (user or company) - Left as comment for now
-                   <Form.Group className="form-group">
-                    <Form.Label>Account Type</Form.Label>
-                    <select
-                      {...register('role')}
-                      className={`form-control ${errors.role ? 'is-invalid' : ''}`}
-                    >
-                      <option value="">Select Account Type</option>
-                      <option value="user">Regular User</option>
-                      <option value="company">Company</option>
-                    </select>
-                    <div className="invalid-feedback">{errors.role?.message}</div>
-                    </Form.Group>
-                  */}
-
-                  <Form.Group className="form-group">
-                    <Form.Label>Password</Form.Label>
-                    <input
-                      type="password"
-                      {...register('password')}
-                      className={`form-control ${errors.password ? 'is-invalid' : ''}`}
-                    />
-                    <div className="invalid-feedback">{errors.password?.message}</div>
-                  </Form.Group>
-                  <Form.Group className="form-group">
-                    <Form.Label>Confirm Password</Form.Label>
-                    <input
-                      type="password"
-                      {...register('confirmPassword')}
-                      className={`form-control ${errors.confirmPassword ? 'is-invalid' : ''}`}
-                    />
-                    <div className="invalid-feedback">{errors.confirmPassword?.message}</div>
-                  </Form.Group>
-                  <Form.Group className="form-group py-3">
-                    <Row>
-                      <Col>
-                        <Button type="submit" className="btn btn-primary">
-                          Register
-                        </Button>
-                      </Col>
-                      <Col>
-                        <Button type="button" onClick={() => reset()} className="btn btn-warning float-right">
-                          Reset
-                        </Button>
-                      </Col>
-                    </Row>
-                  </Form.Group>
-                </Form>
-              </Card.Body>
-              <Card.Footer>
-                Already have an account?
-                <a href="/auth/signin"> Sign in</a>
-              </Card.Footer>
-            </Card>
-          </Col>
-        </Row>
-      </Container>
-    </main>
-  );
-};
+const SignUp = () => (
+  <Container>
+    <Row className="justify-content-center">
+      <Col xs={5} className="text-center py-3">
+        <h2>Welcome to Project Janus</h2>
+        <br />
+        <Card>
+          <Card.Body>
+            <h3 className="text-center">Sign Up</h3>
+            <p className="text-center">Please select your account type.</p>
+            <Row className="justify-content-center">
+              <Col>
+                {/* Send User to the Student Sign Up page */}
+                <Button variant="success" href="/auth/signup/studentsignup">
+                  Student
+                </Button>
+              </Col>
+              <Col>
+                {/* Send User to the Company Sign Up page */}
+                <Button variant="success" href="/auth/signup/companysignup">
+                  Company
+                </Button>
+              </Col>
+            </Row>
+          </Card.Body>
+        </Card>
+      </Col>
+    </Row>
+  </Container>
+);
 
 export default SignUp;
